refactor(shared): add media to hover
