+++ src/app/core/services/theme/theme.ts
@@ -1,27 +1,27 @@
 import { effect, Injectable, signal } from '@angular/core';
-import { THEME } from '../models/theme.enum';
+import { THEME } from '../../models/theme.enum';
 
 @Injectable({
   providedIn: 'root'
 })
 export class Theme {
-  public _theme = signal<THEME.light | THEME.dark>(THEME.light);
+  public theme = signal<THEME.light | THEME.dark>(THEME.light);
 
   constructor() {
     const savedTheme = localStorage.getItem('theme');
 
     if (savedTheme && (savedTheme === THEME.light || savedTheme === THEME.dark)) {
-      this._theme.set(savedTheme);
+      this.theme.set(savedTheme);
     }
 
     effect(() => {
-      localStorage.setItem('theme', this._theme());
-      document.body.className = this._theme();
+      localStorage.setItem('theme', this.theme());
+      document.body.className = this.theme();
     });
   }
 
   public toggleTheme(): void {
-    this._theme.update(currentTheme => (currentTheme === THEME.light ? THEME.dark : THEME.light));
+    this.theme.update(currentTheme => (currentTheme === THEME.light ? THEME.dark : THEME.light));
   }
 
 }
