+++ src/app/shared/header/header.html
@@ -50,7 +50,7 @@
     ><mat-icon>login</mat-icon>
     </a>
 
-    @if (themeService._theme()==="light") {
+    @if (themeService.theme()==="light") {
       <mat-icon class="header__theme" (click)="themeService.toggleTheme()">sunny</mat-icon>
     } @else {
       <mat-icon class="header__theme" (click)="themeService.toggleTheme()">nightlight</mat-icon>
